Add hub_state
