+++ be1-go/.gitignore
@@ -5,3 +5,5 @@ coverage.out
 coverage.html
 configServer1.json
 configServer2.json
+configServer3.json
+configServer4.json
